pyproject.toml
[tool.poetry]
name = "stocks"
version = "0.1.0"
description = ""
authors = ["Viktor Nikolaev <viktornicolaev1995@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
lxml = "^5.2.1"
beautifulsoup4 = "^4.12.3"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
SQLAlchemy = "^2.0.29"
httpx = "^0.27.0"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
alembic = "^1.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
